app/models/preprocess_train.py
"""
NASA Kepler Exoplanet Classification Model
Based on methodologies from Luz et al. (2024) and Malik et al. (2022)
Part of NASA Space Apps 'A World Away' Challenge
"""

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.metrics import (
    confusion_matrix, accuracy_score, precision_score, 
    recall_score, f1_score, roc_auc_score, classification_report
)
from lightgbm import LGBMClassifier
import warnings
warnings.filterwarnings('ignore')

print("=" * 80)
print("PART 1: DATA PREPROCESSING AND CLEANING")
print("=" * 80)

# ============================================================================
# STEP 1: Load Data
# ============================================================================
print("\n[1/7] Loading data from kepler.csv...")
df = pd.read_csv('kepler.csv')
print(f"Original dataset shape: {df.shape}")
print(f"Columns: {df.shape[1]}, Rows: {df.shape[0]}")

# ============================================================================
# STEP 2: Remove Irrelevant Columns
# ============================================================================
print("\n[2/7] Removing irrelevant identifier and non-predictive columns...")
columns_to_drop = [
    'rowid', 'kepid', 'kepoi_name', 'kepler_name', 
    'koi_pdisposition', 'koi_score', 'koi_teq_err1', 'koi_teq_err2'
]
# Only drop columns that exist in the dataframe
columns_to_drop = [col for col in columns_to_drop if col in df.columns]
df = df.drop(columns=columns_to_drop)
print(f"Dropped {len(columns_to_drop)} columns")
print(f"Remaining columns: {df.shape[1]}")

# ============================================================================
# STEP 3: Handle Missing Values
# ============================================================================
print("\n[3/7] Handling missing values...")
print(f"Missing values before imputation: {df.isnull().sum().sum()}")

# Impute missing values with column means for numerical columns
numerical_cols = df.select_dtypes(include=[np.number]).columns
for col in numerical_cols:
    if df[col].isnull().any():
        df[col].fillna(df[col].mean(), inplace=True)

print(f"Missing values after imputation: {df.isnull().sum().sum()}")

# ============================================================================
# STEP 4: Process Target Variable (koi_disposition)
# ============================================================================
print("\n[4/7] Processing target variable (koi_disposition)...")
print(f"Original class distribution:\n{df['koi_disposition'].value_counts()}")

# Filter: Keep only CONFIRMED and CANDIDATE, remove FALSE POSITIVE
df = df[df['koi_disposition'].isin(['CONFIRMED', 'CANDIDATE'])]
print(f"\nDataset shape after filtering: {df.shape}")
print(f"Filtered class distribution:\n{df['koi_disposition'].value_counts()}")

# Convert to binary: CONFIRMED = 0, CANDIDATE = 1
df['koi_disposition'] = df['koi_disposition'].map({'CONFIRMED': 0, 'CANDIDATE': 1})
print(f"\nBinary encoding - CONFIRMED: 0, CANDIDATE: 1")
print(f"Final class distribution:\n{df['koi_disposition'].value_counts()}")

# ============================================================================
# STEP 5: Separate Features and Target
# ============================================================================
print("\n[5/7] Separating features (X) and target (y)...")
X = df.drop('koi_disposition', axis=1)
y = df['koi_disposition']
print(f"Features shape: {X.shape}")
print(f"Target shape: {y.shape}")

# ============================================================================
# STEP 6: Scale Features
# ============================================================================
print("\n[6/7] Scaling features using StandardScaler...")
scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)
X_scaled = pd.DataFrame(X_scaled, columns=X.columns, index=X.index)
print(f"Features scaled successfully")
print(f"Sample statistics (first feature):")
print(f"  Mean: {X_scaled.iloc[:, 0].mean():.6f}")
print(f"  Std: {X_scaled.iloc[:, 0].std():.6f}")

# ============================================================================
# STEP 7: Split Data into Training and Testing Sets
# ============================================================================
print("\n[7/7] Splitting data into training (80%) and testing (20%) sets...")
X_train, X_test, y_train, y_test = train_test_split(
    X_scaled, y, test_size=0.2, random_state=42, stratify=y
)
print(f"X_train shape: {X_train.shape}")
print(f"X_test shape: {X_test.shape}")
print(f"y_train shape: {y_train.shape}")
print(f"y_test shape: {y_test.shape}")
print(f"\nTraining set class distribution:\n{y_train.value_counts()}")
print(f"Testing set class distribution:\n{y_test.value_counts()}")

print("\n" + "=" * 80)
print("DATA PREPROCESSING COMPLETED SUCCESSFULLY")
print("=" * 80)

# ============================================================================
# PART 2: MODEL CREATION, TRAINING, AND EVALUATION
# ============================================================================
print("\n\n" + "=" * 80)
print("PART 2: MODEL CREATION, TRAINING, AND EVALUATION")
print("=" * 80)

# ============================================================================
# STEP 1: Model Selection - LightGBM Classifier
# ============================================================================
print("\n[1/4] Model Selection: LightGBM Classifier (Gradient Boosted Trees)")
print("Following methodology from Malik et al. (2022)")

# ============================================================================
# STEP 2: Hyperparameter Tuning with GridSearchCV
# ============================================================================
print("\n[2/4] Hyperparameter tuning using GridSearchCV with 10-fold CV...")
print("Optimization metric: ROC-AUC")

# Define the parameter grid for LightGBM
param_grid = {
    'n_estimators': [100, 200, 300],
    'learning_rate': [0.01, 0.05, 0.1],
    'num_leaves': [20, 31, 40],
    'max_depth': [5, 7, 10],
    'min_child_samples': [20, 30, 50]
}

# Initialize base LightGBM classifier
lgbm = LGBMClassifier(random_state=42, verbose=-1)

# Set up GridSearchCV with 10-fold cross-validation and ROC-AUC scoring
grid_search = GridSearchCV(
    estimator=lgbm,
    param_grid=param_grid,
    cv=10,
    scoring='roc_auc',
    n_jobs=-1,
    verbose=1
)

print("Starting grid search (this may take several minutes)...")
grid_search.fit(X_train, y_train)

print("\nGrid search completed!")
print(f"Best ROC-AUC score from CV: {grid_search.best_score_:.4f}")
print(f"Best hyperparameters found:")
for param, value in grid_search.best_params_.items():
    print(f"  {param}: {value}")

# Get the best model
best_model = grid_search.best_estimator_

# ============================================================================
# STEP 3: Optimal Decision Threshold Tuning
# ============================================================================
print("\n[3/4] Finding optimal decision threshold to maximize recall...")
print("Strategy: Prioritize finding all potential planets (high recall)")

# Predict probabilities for the positive class (CANDIDATE = 1)
y_proba = best_model.predict_proba(X_test)[:, 1]

# Test different thresholds
thresholds = np.arange(0.1, 0.91, 0.05)
best_threshold = 0.5
best_recall = 0
threshold_results = []

print("\nTesting decision thresholds:")
print(f"{'Threshold':<12} {'Precision':<12} {'Recall':<12} {'F1-Score':<12}")
print("-" * 50)

for threshold in thresholds:
    y_pred_threshold = (y_proba >= threshold).astype(int)
    precision = precision_score(y_test, y_pred_threshold, zero_division=0)
    recall = recall_score(y_test, y_pred_threshold)
    f1 = f1_score(y_test, y_pred_threshold)
    
    threshold_results.append({
        'threshold': threshold,
        'precision': precision,
        'recall': recall,
        'f1': f1
    })
    
    print(f"{threshold:<12.2f} {precision:<12.4f} {recall:<12.4f} {f1:<12.4f}")
    
    # Select threshold with highest recall while maintaining precision > 0.3
    if recall > best_recall and precision > 0.3:
        best_recall = recall
        best_threshold = threshold

print(f"\nOptimal threshold selected: {best_threshold:.2f}")
print(f"This threshold achieves recall of {best_recall:.4f}")

# ============================================================================
# STEP 4: Final Model Evaluation
# ============================================================================
print("\n[4/4] Final model evaluation on test set...")

# Apply optimal threshold to get final predictions
y_pred_final = (y_proba >= best_threshold).astype(int)

# Calculate all performance metrics
conf_matrix = confusion_matrix(y_test, y_pred_final)
accuracy = accuracy_score(y_test, y_pred_final)
precision = precision_score(y_test, y_pred_final)
recall = recall_score(y_test, y_pred_final)
f1 = f1_score(y_test, y_pred_final)
auc = roc_auc_score(y_test, y_proba)

# Print comprehensive performance report
print("\n" + "=" * 80)
print("FINAL MODEL PERFORMANCE REPORT")
print("=" * 80)

print("\n1. CONFUSION MATRIX:")
print(f"\n                Predicted")
print(f"              0 (CONFIRMED)  1 (CANDIDATE)")
print(f"Actual 0      {conf_matrix[0][0]:<15} {conf_matrix[0][1]:<15}")
print(f"       1      {conf_matrix[1][0]:<15} {conf_matrix[1][1]:<15}")

tn, fp, fn, tp = conf_matrix.ravel()
print(f"\nTrue Negatives (TN):  {tn}")
print(f"False Positives (FP): {fp}")
print(f"False Negatives (FN): {fn}")
print(f"True Positives (TP):  {tp}")

print("\n2. CLASSIFICATION METRICS:")
print(f"{'Metric':<20} {'Value':<10}")
print("-" * 30)
print(f"{'Accuracy':<20} {accuracy:.4f}")
print(f"{'Precision':<20} {precision:.4f}")
print(f"{'Recall (Sensitivity)':<20} {recall:.4f}")
print(f"{'F1-Score':<20} {f1:.4f}")
print(f"{'AUC Score':<20} {auc:.4f}")

print("\n3. DETAILED CLASSIFICATION REPORT:")
print(classification_report(
    y_test, y_pred_final, 
    target_names=['CONFIRMED', 'CANDIDATE'],
    digits=4
))

print("\n4. MODEL INTERPRETATION:")
print(f"• The model correctly identified {recall*100:.2f}% of candidate exoplanets")
print(f"• Of all planets predicted as candidates, {precision*100:.2f}% were correct")
print(f"• Overall accuracy: {accuracy*100:.2f}%")
print(f"• AUC score of {auc:.4f} indicates {'excellent' if auc > 0.9 else 'good' if auc > 0.8 else 'fair'} discriminative ability")

print("\n" + "=" * 80)
print("MODEL TRAINING AND EVALUATION COMPLETED SUCCESSFULLY")
print("=" * 80)

print("\n5. SUMMARY:")
print(f"✓ Dataset: {len(df)} KOIs processed")
print(f"✓ Features: {X_train.shape[1]} predictive features")
print(f"✓ Model: LightGBM with optimized hyperparameters")
print(f"✓ Cross-validation: 10-fold, optimized for ROC-AUC")
print(f"✓ Decision threshold: {best_threshold:.2f} (optimized for recall)")
print(f"✓ Final recall: {recall:.4f} - maximizing planet discovery")
print(f"✓ Final precision: {precision:.4f} - maintaining reliability")