requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.7,<3
python-dotenv>=1.0,<2
jinja2>=3.1,<4
numpy
scikit-learn
torch
tensorflow
matplotlib
pandas
hvplot
bokeh
pathlib
seaborn
nbformat
nbconvert
plotly